+++ apps/modernjs-ssr/nested-remote/test.js
@@ -1,0 +1,2 @@
+const serverRender = require('/Users/bytedance/outter/universe/apps/modernjs-ssr/nested-remote/dist/bundles/main.js');
+console.log(33, serverRender);