+++ packages/modernjs/src/runtime/index.ts
@@ -1,0 +1,2 @@
+export * from '@module-federation/enhanced/runtime';
+export { createRemoteSSRComponent } from './createRemoteSSRComponent';