+++ apps/modernjs/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @module-federation/modernjs
 
+## 0.1.39
+
+### Patch Changes
+
+- 85c6a12: fix: support rspack ssr live reload
+  - @module-federation/enhanced@0.3.3
+
+## 0.1.38
+
+### Patch Changes
+
+- Updated dependencies [85ae159]
+  - @module-federation/enhanced@0.3.2
+
+## 0.1.37
+
+### Patch Changes
+
+- @module-federation/enhanced@0.3.1
+
+## 0.1.36
+
+### Patch Changes
+
+- Updated dependencies [fa37cc4]
+  - @module-federation/enhanced@0.3.0
+
+## 0.1.35
+
+### Patch Changes
+
+- @module-federation/enhanced@0.2.8
+
 ## 0.1.34
 
 ### Patch Changes